% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hent_data.R
\name{hent_produksjon_for_kraftverk}
\alias{hent_produksjon_for_kraftverk}
\title{Hent kraftproduksjon for ett selskap}
\usage{
hent_produksjon_for_kraftverk(df, kraftverk = "Aall-Ulefoss")
}
\arguments{
\item{df}{data.frame som må inneholde minst disse kolonnene: Navn og Mid_81_10, der
Navn er navnet på kraftverket, og Mid_81_10 er kraftproduksjonen.}

\item{kraftverk}{Navn på kraftverket man ønsker produksjon til. Kan være en vektor
med flere navn.}
}
\value{
Kraftproduksjon for det eller de angitte kraftverkene
}
\description{
Hent kraftproduksjon for ett selskap
}
