% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hent_data.R
\name{hent_produksjonsdata}
\alias{hent_produksjonsdata}
\title{Hent produksjonsdata fra NVE}
\usage{
hent_produksjonsdata()
}
\value{
data.frame med oversikt over produksjon fra alle kraftverk i Norge.
}
\description{
Funksjon for å hente vannkraftsproduksjon for alle kraftverk i Norge
}
\examples{
# hent_produksjonsdata
}
