% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalkuler.R
\name{kalkuler_kostnader_inntekter}
\alias{kalkuler_kostnader_inntekter}
\title{Title}
\usage{
kalkuler_kostnader_inntekter(prisbane, produksjon, op_cost_kwh = 0.3,
  inv_kwh = 0.2, dep_kwh = 0.2)
}
\arguments{
\item{dep_kwh}{}
}
\value{
returener noe
}
\description{
Title
}
