% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalkuler.R
\name{kalkuler_selskapsverdi}
\alias{kalkuler_selskapsverdi}
\title{Kalkuler nåverdien til selskapet}
\usage{
kalkuler_selskapsverdi(df, wacc)
}
\arguments{
\item{df}{data.frame som har en kolonne kalt kontantflyt. Hver rad representerer
et år, der første rad er det første året.}

\item{wacc}{En numerisk verdi for wacc.}
}
\value{
Returnerer nåverdien for et selskap.
}
\description{
Kalkuler nåverdien til selskapet
}
