% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalkuler.R
\name{kalkuler_wacc}
\alias{kalkuler_wacc}
\title{Kalkulerer wacc med gitte inputverdier}
\usage{
kalkuler_wacc(asset_beta = 0.6, risk_free_rate = 0.02,
  market_risk_premium = 0.05, other_adj = 0.02, debt_premium = 0.025,
  debt_share = 0.5, corp_income_tax = 0.22, inflation = 0.02)
}
\arguments{
\item{inflation}{}
}
\description{
Kalkulerer wacc med gitte inputverdier
}
